requirements.txt
aiohttp==3.6.3
anime==1.1.0
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorlib==0.1.0
colorlog==4.7.2
colour==0.1.5
cryptography==3.4.6
dataclasses-json==0.5.2
dblpy==0.4.0
discord==1.0.1
discord.py==1.6.0
Django==3.1.7
dnspython==2.1.0
fire==0.3.1
futures==3.1.1
gcloud==0.18.3
giphy-client==1.0.0
google==3.0.0
google-api-core==1.26.1
google-api-python-client==1.12.8
google-auth==1.28.0
google-auth-httplib2==0.1.0
Google-Images-Search==1.3.8
googleapis-common-protos==1.53.0
goslate==1.5.1
gsearch==1.6.0
httplib2==0.19.0
idna==2.10
isodate==0.6.0
jwcrypto==0.8
lxml==4.6.3
mal-api==0.4.0
manga==0.1.13
marshmallow==3.10.0
marshmallow-enum==1.5.1
multidict==4.7.6
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
protobuf==3.15.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
PyDictionary==2.0.1
pyfiglet==0.8.post1
pygame==2.0.1
pyjokes==0.6.0
pymongo==3.11.3
PyNaCl==1.4.0
pyparsing==2.4.7
Pyrebase4==4.4.3
python-dateutil==2.8.1
python-jwt==3.3.0
python-mal==0.1.7
python-resize-image==1.1.19
python-youtube==0.6.4
pytube3==9.6.4
pytz==2021.1
randfacts==0.3.0
ratelimiter==1.2.0.post0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2
spotipy==2.16.1
sqlparse==0.4.1
stringcase==1.2.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.3
wget==3.2
wikipedia==1.4.0
yarl==1.5.1
youtube-dl==2021.3.14
